Remove IO code from CompactionJobCommitterOrSendToLambda
